Register types of broadphase into the type registry
