+++ src/broadphase/plugin.js
@@ -3,7 +3,7 @@ import { App, Plugin } from '../app/index.js'
 import { AppSchedule } from '../core/index.js'
 import { PhysicsHitbox } from './components/index.js'
 import { CollisionPairs, Broadphase2D } from './resources/index.js'
-import { getCollisionPairs, updateBroadphase2D } from './systems/index.js'
+import { getCollisionPairs, registerBroadphaseTypes2D, updateBroadphase2D } from './systems/index.js'
 
 export class Broadphase2DPlugin extends Plugin {
 
@@ -27,6 +27,7 @@ export class Broadphase2DPlugin extends Plugin {
   register(app) {
     app
       .registerType(PhysicsHitbox)
+      .registerSystem(AppSchedule.Startup, registerBroadphaseTypes2D)
       .setResource(new Broadphase2D(this.innerBroadphase))
       .setResource(new CollisionPairs())
       .registerSystem(AppSchedule.Update, getCollisionPairs)
